Mobile Sidebar + Images
Added an icon the the top right of the sidebar when on mobile (800px or less viewport). Created a script for the icon. When you click it, the navigation links are toggled to display or not. Edited the styling for the mobile navigation bar and the links. Moved the scripts into their own folder. Edited the About link to have 3 sub pages. Added all the images Brenda sent me to folders organized by page.