Biography
Created the biography page in the about section. Copy pasted Brenda's biography and artist statement. Might move the artist statement if Brenda wants it somewhere else. Set the margins and padding for the containers for normal and mobile. Prelininary styling for <h2> and paragraph text. Added a bottom border to the mobile dropdown menu.