Homepage Nav + Slideshow
Made several changes to the layout of the homepage. Navigation is now a top nav. Added the mobile icon and formatting so now the homepage and sidebar navigations look the same on mobile. Edited the navigation script to account for the homepage. Created the slideshow with temporary items and copied the script from my last website.