Image Gallery Start
Started work on the image gallery in the painting section. Set up the flowers from the sky file since I have the paintings for it. Added all the files to the html. Started styling the gallery images. Added the script to display one image and click through them. However, the script seems to set the height and width of the content to the minimum or something, so trying to align things with "margin: auto;" does not work. Need to find a solution.