Gallery Progress
Figured out how to center the images vertically and horizontally. Adjusted some margins. Added gallery control divs that overlay the image. Left and right work. Center, which switches to thumbnail view, has not yet been implemented. There's an issue with the content or gallery container where the height does not adjust with the browser, so the image starts getting cut on the bottom. Have to figure that out.