Gallery Grid
Fixed the height problem from last push by setting padding on the images. Added a gallery grid when you click on the center control overlay. Added a function to the gallery script to toggle the grid and gallery. Added the opposite toggle to the show images function. The images show in four columns and can be clicked to jump to that image in the gallery view. Still have to add the responsive code.