Responsive grid + Fixes
Added the code to make the gallery image grid responsive. Images will alternate between four, two, and one column depending on viewport size. Fixed some overflow scrollbars on some containers and a display overlook. Fixed something overlaying the sidebar making it so you coudn't click things. Changed the sidebar to be sticky on mobile so it doesn't appear over the images or text. Added a switch statement to the gallery script to only run the gallery function when on desktop. Need to make the images appear properly still.