Responsive Fixes
Tried using grid instead of flex for the image grid but it didn't look as good. Moved a few images around in the Flowers from the Sky gallery. Made gallery images not overlap each other when on mobile. Added a listener to the gallery script to run function depending on screen size. Added a new function to display mobile images correctly when switching from desktop. Added an opacity transition to the grid images on hover. Started image gallery opacity animations when switching images.