Control Panel Controls
Added the image controls to the control panel. Styled the panel and controls. Added lines to the gallery script to show and hide the control panel when the image grid is toggled. Doesn't show properly when changing from mobile to desktop after being in grid mode. Need to add code to remember what mode it was in before switching.