Info + Added Images
Finished the control panel with the information. Added a helper function to the gallery script to change the image meta information when you switch images. Added a profile image to the biography page, and styled it for desktop and mobile. Edited the sidebar to have the dates on the right of the title. Filled the River Dancing gallery with gallery images, onclick grid images, and complete information. Also converted some of the .tif images to .jpg.