Slideshow Animations + More Pages + Colors
Rewrote the home page slideshow script to use the same animations as the gallery script. Split the function into a first image function, for when the script is initially called without animations, and the regular function, for switching the images with animations. Made the slideshow look better on mobile. Added images to all the rest of the painting pages, and meta information to the ones that I have information for. Realized I didn't have the closing html tag for most pages and added it. Renamed "troubled rain" to "troubled waters" in the sidebar and link on all the pages, and the image folder. Changed text colors for the title and sidebar to a slight purple, and the hover and shadow to a lighter purple. I'll see what Brenda thinks.