New Images
Added new images files from Brenda. Added the new River Dancing images in, and rearranged some of the images and files. Made a new page for the Opera House Project and added the images and info to it. Added the new page to the navigation sidebar on all pages. Created a new folder for unused images files. Tried to restructure the grid view to be more responsive, but it didn't work, so I reverted the changes.