More Gallery Pages
Created pages for Oceanus in painting, and Mapping of Memories and Mapping of the Human Heart in drawing. Updated the navigation sidebar on all pages to include the new pages. Added the images to the pages for the gallery and grid, and added info for the images. Corrected the titles for some pages. .tif files don't display, so I will either convert them next push, or find a way to display them.