Video + Resume + Updates
Created a video page with two styled responsive videos from Brenda. Updated the about Resume Page with a link to the resume pdf and two new lines. Moved the Book of Zero series to painting and deleted Mixed Media and its links on all pages. Moved the resume pdf and license to a new files folder.