Link + Publication
Made a page for the external links. Tested each link and added small styling. Made a page for publications, which is a copy of the publications on the resume. Have to ask Brenda if she wants to delete the resume section. Added a fix to the collapsible script in case there is no default id. Removed the accidental highlight from Drawing on the non-collapsible pages. Renamed about container to document container.