Last pages + Favicon
Completed the home page and news page. Added a responsive image to the home page, centered and styled, with a desktop and mobile version. Added a news page with the news Brenda sent, styled with the text container. Added favicon files and meta text to the head section of each page. Compressed the Book of Zero images.